add browser route picker, live sim controls, MissionTrack
